# server/swagger/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Try out API to Common Data Elements"
  version: "0.0.1"
  title: "CDE API"
host: "localhost:3001"
basePath: "/"
schemes:
  - "http"
produces:
  - "application/json"
paths:
  /api/de/{tinyId}:
    get:
      tags:
        - "CDE"
      summary: "CDE element"
      description: "A single data element.\n"
      parameters:
        - name: "tinyId"
          in: "path"
          description: "identifier"
          required: true
          type: "string"
      responses:
        200:
          description: "CDE"
          schema:
            $ref: "#/definitions/De"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "CDE"
  /api/de/{tinyId}/version/{version}:
    get:
      tags:
        - "CDE"
      summary: "CDE element"
      description: "A single data element.\n"
      parameters:
        - name: "tinyId"
          in: "path"
          description: "identifier"
          required: true
          type: "string"
        - name: "version"
          in: "path"
          description: "version"
          required: true
          type: "string"
      responses:
        200:
          description: "CDE"
          schema:
            $ref: "#/definitions/De"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "CDE"
  /api/de/search:
    post:
      tags:
        - "CDE"
      summary: "Data Element document search"
      description: "Full-text search for Data Element documents"
      consumes:
        - application/json
      parameters:
        - in: body
          name: "Search Options"
          schema:
            $ref: "#/definitions/SearchApiOptions"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/SearchDocumentResponseDe"
...
definitions:
  De:
    type: "object"
    properties:
      _id:
        type: "string"
        example: "57112ebf3245ae3033584c5a"
      nihEndorsed:
        type: "boolean"
        example: "true"
      tinyId:
        type: "string"
        example: "QF7dUNkbNEv"
      version:
        type: "string"
        example: "1.0"
      elementType:
        type: "string"
        example: "cde"
      archived:
        type: "boolean"
      views:
        type: "integer"
        format: "int32"
      sources:
        $ref: "#/definitions/Source"
      origin:
        type: "string"
        example: "NCISTD: NCI CRF Harmonization and Standards Initiative"
      imported:
        type: "string"
        format: "date-time"
      created:
        type: "string"
        format: "date-time"
      createdBy:
        $ref: "#/definitions/UserRef"
      updated:
        type: "string"
        format: "date-time"
      updatedBy:
        $ref: "#/definitions/UserRef"
      stewardOrg:
        type: "object"
        properties:
          name:
            type: "string"
            example: "NINDS"
      usedByOrgs:
        type: "array"
        items:
          type: "string"
          example: "CTEP"
      registrationState:
        $ref: "#/definitions/RegistrationState"
      lastMigrationScript:
        type: "string"
        example: "load NINDS on January 15 2020"
      designations:
        $ref: "#/definitions/Designations"
      definitions:
        $ref: "#/definitions/Definitions"
      valueDomain:
        type: "object"
        properties:
          datatype:
            type: "string"
            example: "Date"
          datatypeDate:
            type: "object"
            description: "object appearing only on Date datatypes"
            properties:
              format:
                type: "string"
                example: "YY-MM-DD"
          datatypeDynamicCodeList:
            type: "object"
            description: "object appearing only on Code datatypes"
            properties:
              system:
                type: "string"
                example: "SNOMED"
          datatypeExternallyDefined:
            type: "object"
            description: "object appearing only on Other datatypes"
            properties:
              description:
                type: "string"
                example: "java datatype"
              descriptionFormat:
                type: "string"
                example: "html"
              link:
                type: "string"
          datatypeNumber:
            type: "object"
            description: "object appearing only on Number datatypes"
            properties:
              maxValue:
                type: "number"
                format: "double"
              minValue:
                type: "number"
                format: "double"
              precision:
                type: "number"
                format: "double"
          datatypeText:
            type: "object"
            description: "object appearing only on Text datatypes"
            properties:
              maxLength:
                type: "integer"
                format: "int32"
              minLength:
                type: "integer"
                format: "int32"
              regex:
                type: "string"
                example: ".*:.*:.*"
              rule:
                type: "string"
                example: ""
          datatypeTime:
            type: "object"
            description: "object appearing only on Time datatypes"
            properties:
              format:
                type: "string"
                example: "hh:MM:SS"
          datatypeValueList:
            type: "object"
            description: "object appearing only on Value List datatypes"
            properties:
              datatype:
                type: "string"
                example: "Text"
          definition:
            type: "string"
            example: "to the nearest day"
          identifiers:
            $ref: "#/definitions/Ids"
          ids:
            $ref: "#/definitions/Ids"
          name:
            type: "string"
            example: "Starting Date"
          permissibleValues:
            $ref: "#/definitions/PermissibleValues"
          uom:
            type: "string"
            example: "each"
          vsacOid:
            type: "string"
            example: "1401340134"
      classification:
        type: "array"
        items:
          $ref: "#/definitions/Classification"
      referenceDocuments:
        $ref: "#/definitions/ReferenceDocuments"
      properties:
        $ref: "#/definitions/Properties"
      ids:
        $ref: "#/definitions/Ids"
      attachments:
        $ref: "#/definitions/Attachments"
      comments:
        $ref: "#/definitions/Comments"
      dataSets:
        $ref: "#/definitions/DataSets"
      history:
        $ref: "#/definitions/History"
      derivationRules:
        type: "array"
        items:
          type: "object"
          properties:
            formula:
              type: "string"
              example: "sumAll"
            inputs:
              type: "array"
              items:
                type: "string"
                example: "5h29ApPjjzE"
            name:
              type: "string"
              example: "ALSSS Score"
            outputs:
              type: "array"
              items:
                type: "string"
                example: "10"
            ruleType:
              type: "string"
              example: "score"
...
  Attachments:
    type: "array"
    items:
      type: "object"
      properties:
        comment:
          type: "string"
          example: "Figure 1: Illustration"
        fileid:
          type: "string"
          example: "1"
        filename:
          type: "string"
          example: "1.jpeg"
        filesize:
          type: "integer"
          format: "int64"
        filetype:
          type: "string"
          example: "jpeg"
        isDefault:
          type: "boolean"
        pendingApproval:
          type: "boolean"
        scanned:
          type: "boolean"
        uploadedBy:
          $ref: "#/definitions/UserRef"
        uploadDate:
          type: "string"
          format: "date-time"
  Classification:
    type: "object"
    properties:
      elements:
        type: "array"
        items:
          $ref: "#/definitions/ClassificationElements"
      stewardOrg:
        type: "object"
        properties:
          name:
            type: "string"
            example: "NINDS"
      workingGroup:
        type: boolean
  ClassificationElements:
    type: "object"
    properties:
      elements:
        type: "array"
        items:
          $ref: "#/definitions/ClassificationElements"
      name:
        type: "string"
        example: "Disease"
  Comments:
    type: "array"
    items:
      type: "object"
      properties:
        created:
          type: "string"
          format: "date-time"
        element:
          type: "object"
          properties:
            eltId:
              type: "string"
              example: "views"
            eltType:
              type: "string"
              example: "cde"
        linkedTab:
          type: "string"
          example: "descriptionTab"
        pendingApproval:
          type: "boolean"
        replies:
          type: "array"
          items:
            type: "object"
            properties:
              created:
                type: "string"
                format: "date-time"
              pendingApproval:
                type: "boolean"
              status:
                type: "string"
                example: "resolved"
              text:
                type: "string"
                example: "update value"
              user:
                type: "string"
                example: "jdoe"
              usename:
                type: "string"
                example: "jdoe"
        status:
          type: "string"
          example: "resolved"
        text:
          type: "string"
          example: "update value"
        user:
          type: "string"
          example: "jdoe"
        usename:
          type: "string"
          example: "jdoe"
  CurationStatus:
    type: "string"
    enum:
      - "Qualified"
      - "Standard"
      - "Preferred Standard"
  DataSets:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "string"
          example: "15232523"
        notes:
          type: "string"
          example: "trial 152"
        source:
          type: "string"
          example: "NINDS"
        studyUri:
          type: "string"
          example: "http://www.ninds.org/15232523"
  Datatype:
    type: string
    enum:
      - "Date"
      - "Dynamic Code List"
      - "Externally Defined"
      - "File"
      - "Geo Location"
      - "Number"
      - "Text"
      - "Time"
      - "Value List"
  History:
    type: "array"
    items:
      type: "string"
      example: "57112ebf3245ae3033584c5a"
  Ids:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "string"
          example: "Qd4_a8FZqrX"
        source:
          type: "string"
          example: "NLM"
        version:
          type: "string"
          example: "3"
...
  Designations:
    type: "array"
    items:
      type: "object"
      properties:
        designation:
          type: "string"
          example: "Title Name"
        source:
          type: "string"
          example: "NINDS"
        tags:
          type: "array"
          items:
            type: "object"
            properties:
              tag:
                type: "string"
                example: "Health"
  Definitions:
    type: "array"
    items:
      type: "object"
      properties:
        definition:
          type: "string"
          example: "This is a description of purpose"
        definitionFormat:
          type: "string"
          example: "html"
        source:
          type: "string"
          example: "NINDS"
        tags:
          type: "array"
          items:
            type: "object"
            properties:
              tag:
                type: "string"
                example: "Health"
  PermissibleValues:
    type: "array"
    items:
      type: "object"
      properties:
        codeSystemName:
          type: "string"
          example: "NCI Thesaurus"
        codeSystemVersion:
          type: "string"
          example: "1.0"
        permissibleValue:
          type: "string"
          example: "yes"
        valueMeaningCode:
          type: "string"
          example: "C46110"
        valueMeaningDefinition:
          type: "string"
          example: "affirmative declaration"
        valueMeaningName:
          type: "string"
          example: "Yes"
  Properties:
    type: "array"
    items:
      type: "object"
      properties:
        key:
          type: "string"
          example: "custom1"
        source:
          example: "NINDS"
          type: "string"
        value:
          type: "string"
          example: "10"
        valueFormat:
          type: "string"
          example: "html"
  ReferenceDocuments:
    type: "array"
    items:
      type: "object"
      properties:
        document:
          type: "string"
          example: "rd1"
        docType:
          type: "string"
          example: "rd1"
        languageCode:
          type: "string"
          example: "EN-US"
        providerOrg:
          type: "string"
          example: "NINDS"
        referenceDocumentId:
          type: "string"
          example: "ref1"
        source:
          type: "string"
          example: "NINDS"
        text:
          type: "string"
          example: "rd1"
        title:
          type: "string"
          example: "rd1"
        uri:
          type: "string"
          example: "http://www.ninds.com/125123u"
  RegistrationState:
    type: "object"
    properties:
      administrativeNote:
        type: "string"
        example: "approved"
      administrativeStatus:
        type: "string"
        example: "Nominal"
      effectiveDate:
        type: "string"
        format: "date-time"
      registrationStatus:
        type: "string"
        example: "Qualified"
      replacedBy:
        type: "object"
        properties:
          tinyId:
            type: "string"
            example: "Qd4_a8FZqrX"
      unresolvedIssue:
        type: "string"
        example: "none"
      untilDate:
        type: "string"
        format: "date-time"
  SearchDocumentResponseDe:
    type: object
    properties:
      resultsTotal:
        type: integer
        format: int32
      resultsRetrieved:
        type: integer
        format: int32
      from:
        type: integer
        format: int32
      docs:
        type: "array"
        items:
          $ref: "#/definitions/De"
    required: [ "resultsTotal", "resultsRetrieved", "from", "docs" ]
...
  SearchApiOptions:
    type: object
    properties:
      nihEndorsed:
        description: "Endorsed by NIH"
        type: boolean
      excludeAllOrgs:
        description: "Exclude documents with more than 1 organization"
        type: boolean
      excludeOrgs:
        description: "Exclude document from specific organizations"
        type: array
        items:
          type: string
      page:
        description: "Specify page of result set, starting with 1"
        type: integer
      resultPerPage:
        description: "Number of results to retrieve at one time, maximum 100"
        type: integer
      searchTerm:
        description: "Search Query, see examples at https://cde.nlm.nih.gov/guides"
        type: string
      selectedDatatypes:
        description: "Data Elements with specific datatype"
        type: array
        items:
          $ref: "#/definitions/Datatype"
      selectedElements:
        description: "Classification"
        type: array
        items:
          type: string
      selectedElementsAlt:
        description: "Second Classification"
        type: array
        items:
          type: string
      selectedOrg:
        description: "Classifying Organization, not to be confused with steward"
        type: string
      selectedOrgAlt:
        description: "Second Classifying Organization, not to be confused with steward"
        type: string
      selectedStatuses:
        description: "Registration Status"
        type: array
        items:
          $ref: "#/definitions/CurationStatus"
        uniqueItems: true
  Source:
    type: "object"
    properties:
      sourceName:
        type: "string"
        example: "caDSR"
      created:
        type: "string"
        format: "date"
      updated:
        type: "string"
        format: "date"
      registrationStatus:
        type: "string"
        example: "Qualified"
      datatype:
        type: "string"
        example: "Number"
      copyright:
        type: "object"
        properties:
          value:
            type: "string"
            example: "MIT license"
          valueFormat:
            type: "string"
            example: "html"
  UserRef:
    type: "object"
    properties:
      username:
        type: "string"
        example: "jdoe"
      userId:
        type: "string"
        example: 'ObjectId("5711293936f82f59878e9c27")'
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      fields:
        type: "string"